Move SlotState tests to SlotStateTest.
